% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimMET.R
\name{apsimMET}
\alias{apsimMET}
\title{Convert PPD data to APSIM .met file}
\usage{
apsimMET(id, start, end, username, password)
}
\arguments{
\item{id}{Weather station ID.}

\item{start}{Start date (String YYYYmmdd).}

\item{end}{End date (String YYYYmmdd).}

\item{username}{SILO username (available from https://www.longpaddock.qld.gov.au/silo/).}

\item{password}{SILO password (available form https://www.longpaddock.qld.gov.au/silo/).}
}
\value{
Data frame containing daily weather data.
}
\description{
Download PPD data for a station and return as a data frame
}
\author{
Fiona Evans
}
