% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callington_met.R
\name{callington_met}
\alias{callington_met}
\title{Callington (SA) met producer}
\usage{
callington_met(UserName, PassWord, startdate)
}
\arguments{
\item{UserName}{as "QLDUni" by default}

\item{PassWord}{as "Q1dUn!" by default}

\item{startdate}{as "17/May/2021" by default, alternatively add "%20HH:MM:SS"}
}
\value{
output
}
\description{
Designed for Callington dataset (up to date) from Wildeye API
}
\author{
patrickaxe
}
