% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimMET.R
\name{dpird2apsimMET}
\alias{dpird2apsimMET}
\title{Convert downloaded DPIRD data to APSIM .met file}
\usage{
dpird2apsimMET(id, weather, stations, fileName = "weather.met")
}
\arguments{
\item{id}{Weather station ID.}

\item{weather}{data frame containing DPIRD weather data downloaded using getDPIRDdaily().}

\item{stations}{list of DPIRD weather stations downloaded using getDPIRDstations().}

\item{fileName}{name of .met file to write output to.}
}
\value{
Data frame containing daily weather data.
}
\description{
Convert downloaded DPIRD data to APSIM .met file
}
\author{
Fiona Evans
}
