% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met.R
\name{get_met}
\alias{get_met}
\title{https://www.longpaddock.qld.gov.au/cgi-bin/silo/PatchedPointDataset.php?format=near&station=15540&radius=100000}
\usage{
get_met(lat, lon, startdate, enddate)
}
\arguments{
\item{lat}{is the latitude of a location}

\item{lon}{is the longtitude of a location}

\item{startdate}{date format in YYYYmmdd}

\item{enddate}{date format in YYYmmdd}
}
\description{
https://www.longpaddock.qld.gov.au/cgi-bin/silo/PatchedPointDataset.php?format=near&station=15540&radius=100000
}
\author{
patrickaxe
}
