% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDPIRD.R
\name{getDPIRDdaily}
\alias{getDPIRDdaily}
\title{Download DPIRD daily weather data for one year (may contain missing values)}
\usage{
getDPIRDdaily(id, year, username, password)
}
\arguments{
\item{id}{Weather station ID.}

\item{year}{Year (numeric YYYY).}

\item{apiKey}{API key (available from https://www.agric.wa.gov.au/web-apis).}
}
\value{
Data frame containing daily weather data.
}
\description{
Download DPIRD daily weather data for one year (may contain missing values).
}
\author{
Fiona Evans
}
