% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDPIRD.R
\name{getDPIRDhourly}
\alias{getDPIRDhourly}
\title{Download DPIRD hourly weather data for a period of one year (may contain missing values)}
\usage{
getDPIRDhourly(id, start, end, apiKey)
}
\arguments{
\item{id}{Weather station ID.}

\item{start}{Start date (Date).}

\item{end}{End date (Date).}

\item{apiKey}{API key (available from https://www.agric.wa.gov.au/web-apis).}
}
\value{
Data frame containing daily weather data.
}
\description{
Download DPIRD daily weather data for a period of one year (may contain missing values)
}
\author{
Fiona Evans
}
