% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDPIRD.R
\name{getDPIRDhourlyByDay}
\alias{getDPIRDhourlyByDay}
\title{Download DPIRD hourly weather data for a period of one day (may contain missing values)}
\usage{
getDPIRDhourlyByDay(id, year, month, day, apiKey)
}
\arguments{
\item{id}{Weather station ID.}

\item{year}{Year (numeric YYYY).}

\item{month}{Month (numeric m).}

\item{day}{Day (numeric d).}

\item{apiKey}{API key (available from https://www.agric.wa.gov.au/web-apis).}
}
\value{
Data frame containing daily weather data.
}
\description{
Download DPIRD daily weather data for a period of one day (may contain missing values)
}
\author{
Fiona Evans
}
