% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDPIRD.R
\name{getDPIRDstations}
\alias{getDPIRDstations}
\title{Download list of DPIRD automatic weather stations}
\usage{
getDPIRDstations(apiKey)
}
\arguments{
\item{apiKey}{API key (available from https://www.agric.wa.gov.au/web-apis).}
}
\value{
Data frame containing list oif weather stations and metadata..
}
\description{
Download list of DPIRD automatic weather stations
}
\author{
Fiona Evans
}
