% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPPD.R
\name{getPPD.apiKey}
\alias{getPPD.apiKey}
\title{Download PPD data}
\usage{
getPPD.apiKey(id, start, end, silo.apiKey)
}
\arguments{
\item{id}{Weather station ID.}

\item{start}{Start date (String YYYYmmdd).}

\item{end}{End date (String YYYYmmdd).}

\item{silo.apiKey}{SILO API key (available from https://silo.longpaddock.qld.gov.au/).}
}
\value{
Data frame containing daily weather data.
}
\description{
Download PPD data for a station and return as a data frame.
Previous vesion of SILO required an apiKey
}
\author{
Fiona Evans
}
