% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSCIENCE.R
\name{getRTD}
\alias{getRTD}
\title{Download  current 'rainfall to date' data using DPIRD SCIENCE API}
\usage{
getRTD(stations.id, apiKey)
}
\arguments{
\item{apiKey}{API key (see https://www.agric.wa.gov.au/science-api-20).}

\item{id}{Weather station ID.}
}
\value{
List of data frames.
}
\description{
Download current 'rainfall to date' data using DPIRD SCIENCE API
}
\author{
Fiona Evans
}
