% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSCIENCE.R
\name{getSCIENCEstations}
\alias{getSCIENCEstations}
\title{Download list of DPIRD SCIENCE weather stations}
\usage{
getSCIENCEstations(apiKey)
}
\arguments{
\item{apiKey}{API key (see https://www.agric.wa.gov.au/science-api-20).}
}
\value{
Data frame containing list of weather stations and metadata..
}
\description{
Download list of DPIRD SCIENCE weather stations - contains Patched Poitn and DPIRD
automatic weather stations
}
\author{
Fiona Evans
}
