% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSCIENCE.R
\name{getSoilWater}
\alias{getSoilWater}
\title{Download  'soil water' data using DPIRD SCIENCE API}
\usage{
getSoilWater(station.id, startDate, endDate,
  soilType = "shallow-sandy-duplex", faoInitialisationDays = 15,
  faoDevelopmentDays = 20, faoMidSeasonDays = 80, faoLateSeasonDays = 45,
  faoInitialisationCropCoefficient = 0.05,
  faoDevelopmentCropCoefficient = 0.1, faoMidSeasonCropCoefficient = 0.45,
  faoLateSeasonCropCoefficient = 0.2, faoBreakOfSeason3Days25April = 15,
  faoBreakOfSeason3Days5June = 5, apiKey)
}
\arguments{
\item{station.id}{Station id}

\item{startDate}{Summer start date (sw is initialised at zero); string "YYYY-mm-dd"}

\item{endDate}{End date; string "YYYY-mm-dd"}

\item{soilType}{Soil type; string element of c("gravel, "shallow-soil", "sand", "sandy-earth", "shallow-sandy-duplex", "deep-sand-duplex", "shallow-loamy-duplex", "deep-loamy-"duplex", "loamy-earth", "clay")}

\item{faoInitialisationDays}{number of days in crop initialisation, used in calculation of crop evapotranspiration}

\item{faoDevelopmentDays}{number of days in crop development period, used in calculation of crop evapotranspiration}

\item{faoMidSeasonDays}{number of days in crop mid-season period, used in calculation of crop evapotranspiration}

\item{faoLateSeasonDays}{number of days in crop late-season period, used in calculation of crop evapotranspiration}

\item{faoInitialisationCropCoefficient}{proportion of potential evaporation used by crop during initialisation period}

\item{faoDevelopmentCropCoefficient}{proportion of potential evaporation used by crop during development period}

\item{faoMidSeasonCropCoefficient}{proportion of potential evaporation used by crop during mid-season period}

\item{faoLateSeasonCropCoefficient}{proportion of potential evaporation used by crop during late-season period}

\item{faoBreakOfSeason3Days25April}{mm of rainfall after April 25th that triggers the break of season}

\item{faoBreakOfSeason3Days5June}{mm of rainfall after June 5th that triggers the break of season}

\item{apiKey}{API key (see https://www.agric.wa.gov.au/science-api-20).}
}
\value{
List of data frames.
}
\description{
Download 'soil water' data using DPIRD SCIENCE API.
See https://www.agric.wa.gov.au/climate-weather/soil-water-tool
for details on the soil water model used.
}
\author{
Fiona Evans
}
