% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestStations.R
\name{nearestStations}
\alias{nearestStations}
\title{Find the nearest stations to a given longitude and latitude}
\usage{
nearestStations(lon, lat)
}
\arguments{
\item{lon}{Longitude.}

\item{lat}{Latitude.}
}
\value{
Returns a vector of station ids, in order of increasing distance.
}
\description{
Find the nearest stations to a given longitude and latitude
}
\author{
Fiona Evans
}
