% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimMET.R
\name{ppd2apsimMET}
\alias{ppd2apsimMET}
\title{Convert downloaded PPD data to APSIM .met file}
\usage{
ppd2apsimMET(id, weather, fileName = "weather.met")
}
\arguments{
\item{id}{Weather station ID.}

\item{weather}{data frame containing PPD weather data downloaded using getPPD().}

\item{fileName}{name of .met file to write output to.}
}
\value{
Data frame containing daily weather data.
}
\description{
Convert downloaded PPD data to APSIM .met file
}
\author{
Fiona Evans
}
