% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/westdale_met.R
\name{westdale_met}
\alias{westdale_met}
\title{West Dale (WA) met producer}
\usage{
westdale_met(get)
}
\arguments{
\item{type}{in "get" to execute the function}
}
\description{
Designed for West Dale (WA) dataset (up to date) from avachat.info
}
\author{
patrickaxe
}
