% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDPIRD.R
\name{fillDPIRDdaily}
\alias{fillDPIRDdaily}
\title{Fill missing DPIRD daily weather data (for one year) with data from nearest PPD station.}
\usage{
fillDPIRDdaily(id, year, weather, dpird.apiKey, silo.apiKey,
  start = NULL, end = NULL)
}
\arguments{
\item{id}{Weather station ID.}

\item{year}{Year.}

\item{weather}{data frame containing DPIRD weather data downloaded using getDPIRDdaily().}

\item{silo.apiKey}{SILO API key (available from https://silo.longpaddock.qld.gov.au/).}

\item{start}{Optional start date (class Date) from which to fill}

\item{end}{Optional end date (class Date) to fill to - useful if you want to fill 
the entire year. If start and end are omitted, the start and end dates in the input
weather data frame are used.}
}
\value{
Data frame containing filled daily weather data.
}
\description{
Fill missing DPIRD daily weather data (for one year) with data from nearest PPD station.
}
\author{
Fiona Evans
}
