% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClimatology.R
\name{getClimatology}
\alias{getClimatology}
\title{Download PPD data and store climatology}
\usage{
getClimatology(id, start.year, end.year, username = "MUEVANS",
  password = "FIONA6151")
}
\arguments{
\item{id}{Weather station ID.}

\item{start.year}{Start year.}

\item{end.year}{End year.}

\item{username}{SILO username (available from https://www.longpaddock.qld.gov.au/silo/).}

\item{password}{SILO password (available form https://www.longpaddock.qld.gov.au/silo/).}
}
\value{
Data frame containing daily weather data (leap year days omitted).
}
\description{
Download PPD data for a station and store climatology as a data frame where the 3rd dimension index by year
}
\author{
Fiona Evans
}
