% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDPIRD.R
\name{getDPIRDdaily}
\alias{getDPIRDdaily}
\title{Download DPIRD daily weather data (may contain missing values)}
\usage{
getDPIRDdaily(id, start, end, apiKey)
}
\arguments{
\item{id}{Weather station ID.}

\item{start}{Start date (String YYYYmmdd).}

\item{end}{End date (String YYYYmmdd).}

\item{apiKey}{API key (available from https://www.agric.wa.gov.au/web-apis).}
}
\value{
Data frame containing daily weather data.
}
\description{
Download DPIRD daily weather data (may contain missing values)
}
\author{
Fiona Evans
}
